Logic: Add comparison with Integer
Test `test_my_design.my_first_test` and `test_my_design.my_second_test` are
doing comparison with a bit and an integer.
In order to make LogicArray behave similarly to BinaryValue we also need to
implement comparison between Bit and int.

I also add a doctest in logic_array.py to make sure we don't break it when
refactoring LogicArray to improve performance. (It is not in logic.py because we
may remove it when refactoring)